Sound, Fish selling, bug fixing.
Changed the start menu quit button to a web build friendly quit button, (when clicked on web will just redirect to start so doesnt do anything yet doesnt give error). Added fitting fonts, sound effects and music. Setup everything music and sound related I had as well as the fonts. Music changes automatically and correctly. Added the value of the fish to the inventory display, selected fishes can now be sold through inventory and the amount will be added to the player their wallet. Only bug I found was that once a fish was sold the game would not fill that slot again but just kept moving forward in slot number. Bug is now fixed.

+++ FishIsland2D/Assets/Scripts/Fishes/Scriptable Objects/FishData.cs
@@ -9,8 +9,8 @@ public class FishData : ScriptableObject
     public string fishTier; //Legendary, Epic, Rare, Uncommon, Common
     [Range(0f, 100f)] public float catchChance = 10f; //Higher = more likely to be caught
 
-    public float fishMaxSize;
-    public float fishMinSize;
+    public int fishMaxSize;
+    public int fishMinSize;
 
-    public float sellPrice;
+    public int sellPrice;
 }
